+++ R/predictions.R
@@ -48,6 +48,8 @@ select_prediction_method <- function(model.class,
     fitfram <- get_predictions_geeglm(model, expanded_frame, ...)
   } else if (model.class == "gamlss") {
     fitfram <- get_predictions_gamlss(model, expanded_frame, ci.lvl, linv, terms, model.class, typical, condition, ...)
+  } else if (model.class == "bamlss") {
+    fitfram <- get_predictions_bamlss(model, expanded_frame, linv, ...)
   } else if (model.class == "gam") {
     fitfram <- get_predictions_gam(model, expanded_frame, ci.lvl, linv, type, ...)
   } else if (model.class == "Gam") {